Now sends email after set interval passes
