+++ server/app.js
@@ -9,6 +9,8 @@ import mongoose from 'mongoose';
 mongoose.Promise = require('bluebird');
 import config from './config/environment';
 import http from 'http';
+//EmailReminder
+var emailReminder = require('./api/emailReminder/emailReminder.js');
 
 // Connect to MongoDB
 mongoose.connect(config.mongo.uri, config.mongo.options);
@@ -31,6 +33,9 @@ function startServer() {
   app.angularFullstack = server.listen(config.port, config.ip, function() {
     console.log('Express server listening on %d, in %s mode', config.port, app.get('env'));
   });
+  //Start the CronJob
+  emailReminder.loop_job.start();
+  console.log('loopJob status', emailReminder.loop_job.running); // loopJob status true if running
 }
 
 setImmediate(startServer);
